\name{solomon}
\alias{solomon}
\alias{SOLOMON}
\alias{ALLELES}
\alias{ALLELEFREQS}
\alias{DADS}
\alias{MOMS}
\alias{NALLELES}
\alias{NLOCI}
\alias{OFFSPRING}
\alias{adults}
\alias{loci_number}
\alias{offspring}
\alias{solomon.env}
\title{Solomon}
\description{
function to launch SOLOMON graphical user interface for Bayesian Parentage Analysis. See http://www.science.oregonstate.edu/~christim/SOLOMON.html for a detailed user manualto the graphical user interface.}
\usage{
solomon()
}
\details{
launches SOLOMON GUI
}
\references{
Christie et al (2013) Bioinformatics in press
see http://www.science.oregonstate.edu/~christim/SOLOMON.html for a full user guide and FAQs
}
\author{
Mark Christie
}
\examples{\dontrun{solomon()
}}
\keyword{Parentage}

